+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/ActiveContractsService.scala
@@ -1,0 +1,15 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index.v1
+
+import com.digitalasset.ledger.api.domain.TransactionFilter
+
+import scala.concurrent.Future
+
+trait ActiveContractsService {
+
+  def getActiveContractSetSnapshot(
+      filter: TransactionFilter
+  ): Future[ActiveContractSetSnapshot]
+}
