+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/TimeService.scala
@@ -1,0 +1,12 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index.v1
+
+import akka.NotUsed
+import akka.stream.scaladsl.Source
+import com.digitalasset.daml.lf.data.Time
+
+trait TimeService {
+  def getLedgerRecordTimeStream(): Source[Time.Timestamp, NotUsed]
+}
