+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/TransactionsService.scala
@@ -9,6 +9,8 @@ import com.daml.ledger.participant.state.v1.Offset
 import com.digitalasset.daml.lf.transaction.BlindingInfo
 import com.digitalasset.ledger.api.domain.TransactionFilter
 
+import scala.concurrent.Future
+
 trait TransactionsService {
   // FIXME(JM): Cleaner name/types for this
   // FIXME(JM): Fold BlindingInfo into TransactionAccepted, or introduce
@@ -18,6 +20,10 @@ trait TransactionsService {
       endAt: Option[Offset],
       filter: TransactionFilter
   ): Source[(Offset, (TransactionAccepted, BlindingInfo)), NotUsed]
+
+  def getLedgerBeginning(): Future[Offset]
+
+  def getLedgerEnd(): Future[Offset]
 
   /*
   def getTransactionById(
